chore: refactor to SgidLoginSection
